Refactor admin dashboard to Astro with small React button islands
Replace the monolithic AdminDashboard React component with Astro
markup for the page structure (headings, post list, links) and two
small React islands (LogoutButton, DeletePostButton) only for the
click handlers that need client-side JS.
